+++ playground/components/TheSidebar.vue
@@ -9,6 +9,7 @@ interface Props {
 
 interface Emits {
   (e: 'update:activeExample', id: string): void;
+  (e: 'clickSourceCode'): void;
 }
 
 const props = withDefaults(defineProps<Props>(), { examples: () => [] });
@@ -27,6 +28,7 @@ function handleItemClick(example: Example) {
       :text="example.title"
       :active="props.activeExample === example.id"
       @click="() => handleItemClick(example)"
+      @clickSourceCode="$emit('clickSourceCode')"
     />
   </ul>
 </template>