+++ playground/examples/index.ts
@@ -3,9 +3,4 @@ import keyboardExample from './keyboard';
 
 export default [
   keyboardExample, //
-  {
-    ...keyboardExample,
-    id: 'momentosotsotsos',
-    title: 'Keyboard Example 2',
-  },
 ] as Example[];